my_project/requirements.txt
Django==3.2.10
djangorestframework==3.12.4
gunicorn==20.1.0
